+++ src/state/annotation_utils.test.ts
@@ -3,15 +3,24 @@ import {
   AnnotationDomainTypes,
   AnnotationSpec,
   AxisSpec,
+  LineAnnotationSpec,
   Position,
   RectAnnotationSpec,
   Rotation,
-  LineAnnotationSpec,
 } from '../lib/series/specs';
 import { DEFAULT_ANNOTATION_LINE_STYLE } from '../lib/themes/theme';
 import { Dimensions } from '../lib/utils/dimensions';
-import { AnnotationId, AxisId, getAnnotationId, getAxisId, getGroupId, GroupId } from '../lib/utils/ids';
-import { createContinuousScale, createOrdinalScale, Scale, ScaleType } from '../lib/utils/scales/scales';
+import {
+  AnnotationId,
+  AxisId,
+  getAnnotationId,
+  getAxisId,
+  getGroupId,
+  GroupId,
+} from '../lib/utils/ids';
+import { ScaleBand } from '../lib/utils/scales/scale_band';
+import { ScaleContinuous } from '../lib/utils/scales/scale_continuous';
+import { Scale, ScaleType } from '../lib/utils/scales/scales';
 import {
   AnnotationLinePosition,
   AnnotationLineProps,
@@ -45,10 +54,13 @@ describe('annotation utils', () => {
   const maxRange = 100;
 
   const continuousData = [0, 10];
-  const continuousScale = createContinuousScale(ScaleType.Linear, continuousData, minRange, maxRange);
+  const continuousScale = new ScaleContinuous(ScaleType.Linear, continuousData, [
+    minRange,
+    maxRange,
+  ]);
 
   const ordinalData = ['a', 'b', 'c', 'd', 'a', 'b', 'c'];
-  const ordinalScale = createOrdinalScale(ordinalData, minRange, maxRange);
+  const ordinalScale = new ScaleBand(ordinalData, [minRange, maxRange]);
 
   const chartDimensions: Dimensions = {
     width: 10,
@@ -192,11 +204,13 @@ describe('annotation utils', () => {
       xScale,
       Position.Left,
     );
-    const expectedDimensions = [{
-      position: [DEFAULT_LINE_OVERFLOW, 20, 10, 20],
-      details: { detailsText: 'foo', headerText: '2' },
-      tooltipLinePosition: [0, 20, 10, 20],
-    }];
+    const expectedDimensions = [
+      {
+        position: [DEFAULT_LINE_OVERFLOW, 20, 10, 20],
+        details: { detailsText: 'foo', headerText: '2' },
+        tooltipLinePosition: [0, 20, 10, 20],
+      },
+    ];
     expect(dimensions).toEqual(expectedDimensions);
   });
 
@@ -225,11 +239,13 @@ describe('annotation utils', () => {
       xScale,
       Position.Right,
     );
-    const expectedDimensions = [{
-      position: [0, 20, 10, 20],
-      details: { detailsText: 'foo', headerText: '2' },
-      tooltipLinePosition: [0, 20, 10, 20],
-    }];
+    const expectedDimensions = [
+      {
+        position: [0, 20, 10, 20],
+        details: { detailsText: 'foo', headerText: '2' },
+        tooltipLinePosition: [0, 20, 10, 20],
+      },
+    ];
     expect(dimensions).toEqual(expectedDimensions);
   });
 
@@ -258,11 +274,13 @@ describe('annotation utils', () => {
       xScale,
       Position.Left,
     );
-    const expectedDimensions = [{
-      position: [20, 0, 20, 20 + DEFAULT_LINE_OVERFLOW],
-      details: { detailsText: 'foo', headerText: '2' },
-      tooltipLinePosition: [20, 0, 20, 20],
-    }];
+    const expectedDimensions = [
+      {
+        position: [20, 0, 20, 20 + DEFAULT_LINE_OVERFLOW],
+        details: { detailsText: 'foo', headerText: '2' },
+        tooltipLinePosition: [20, 0, 20, 20],
+      },
+    ];
     expect(dimensions).toEqual(expectedDimensions);
   });
 
@@ -315,11 +333,13 @@ describe('annotation utils', () => {
       xScale,
       Position.Left,
     );
-    const expectedDimensions = [{
-      position: [12.5, -DEFAULT_LINE_OVERFLOW, 12.5, 20],
-      details: { detailsText: 'foo', headerText: 'a' },
-      tooltipLinePosition: [12.5, 0, 12.5, 20],
-    }];
+    const expectedDimensions = [
+      {
+        position: [12.5, -DEFAULT_LINE_OVERFLOW, 12.5, 20],
+        details: { detailsText: 'foo', headerText: 'a' },
+        tooltipLinePosition: [12.5, 0, 12.5, 20],
+      },
+    ];
     expect(dimensions).toEqual(expectedDimensions);
   });
 
@@ -346,11 +366,13 @@ describe('annotation utils', () => {
       xScale,
       Position.Top,
     );
-    const expectedDimensions = [{
-      position: [20, -DEFAULT_LINE_OVERFLOW, 20, 20],
-      details: { detailsText: 'foo', headerText: '2' },
-      tooltipLinePosition: [20, 0, 20, 20],
-    }];
+    const expectedDimensions = [
+      {
+        position: [20, -DEFAULT_LINE_OVERFLOW, 20, 20],
+        details: { detailsText: 'foo', headerText: '2' },
+        tooltipLinePosition: [20, 0, 20, 20],
+      },
+    ];
     expect(dimensions).toEqual(expectedDimensions);
   });
 
@@ -377,11 +399,13 @@ describe('annotation utils', () => {
       xScale,
       Position.Bottom,
     );
-    const expectedDimensions = [{
-      position: [20, DEFAULT_LINE_OVERFLOW, 20, 20],
-      details: { detailsText: 'foo', headerText: '2' },
-      tooltipLinePosition: [20, 0, 20, 20],
-    }];
+    const expectedDimensions = [
+      {
+        position: [20, DEFAULT_LINE_OVERFLOW, 20, 20],
+        details: { detailsText: 'foo', headerText: '2' },
+        tooltipLinePosition: [20, 0, 20, 20],
+      },
+    ];
     expect(dimensions).toEqual(expectedDimensions);
   });
 
@@ -409,11 +433,13 @@ describe('annotation utils', () => {
       xScale,
       Position.Left,
     );
-    const expectedDimensions = [{
-      position: [-DEFAULT_LINE_OVERFLOW, 12.5, 10, 12.5],
-      details: { detailsText: 'foo', headerText: 'a' },
-      tooltipLinePosition: [0, 12.5, 10, 12.5],
-    }];
+    const expectedDimensions = [
+      {
+        position: [-DEFAULT_LINE_OVERFLOW, 12.5, 10, 12.5],
+        details: { detailsText: 'foo', headerText: 'a' },
+        tooltipLinePosition: [0, 12.5, 10, 12.5],
+      },
+    ];
     expect(dimensions).toEqual(expectedDimensions);
   });
 
@@ -441,11 +467,13 @@ describe('annotation utils', () => {
       xScale,
       Position.Left,
     );
-    const expectedDimensions = [{
-      position: [-DEFAULT_LINE_OVERFLOW, 20, 10, 20],
-      details: { detailsText: 'foo', headerText: '2' },
-      tooltipLinePosition: [0, 20, 10, 20],
-    }];
+    const expectedDimensions = [
+      {
+        position: [-DEFAULT_LINE_OVERFLOW, 20, 10, 20],
+        details: { detailsText: 'foo', headerText: '2' },
+        tooltipLinePosition: [0, 20, 10, 20],
+      },
+    ];
     expect(dimensions).toEqual(expectedDimensions);
   });
 
@@ -478,7 +506,8 @@ describe('annotation utils', () => {
         position: [-DEFAULT_LINE_OVERFLOW, 0, 10, 0],
         details: { detailsText: 'foo', headerText: '2' },
         tooltipLinePosition: [0, 0, 10, 0],
-      }];
+      },
+      ];
       expect(dimensions).toEqual(expectedDimensions);
     });
 
@@ -506,11 +535,13 @@ describe('annotation utils', () => {
       xScale,
       Position.Top,
     );
-    const expectedDimensions = [{
-      position: [-10, -DEFAULT_LINE_OVERFLOW, -10, 20],
-      details: { detailsText: 'foo', headerText: '2' },
-      tooltipLinePosition: [-10, 0, -10, 20],
-    }];
+    const expectedDimensions = [
+      {
+        position: [-10, -DEFAULT_LINE_OVERFLOW, -10, 20],
+        details: { detailsText: 'foo', headerText: '2' },
+        tooltipLinePosition: [-10, 0, -10, 20],
+      },
+    ];
     expect(dimensions).toEqual(expectedDimensions);
   });
 
@@ -542,7 +573,8 @@ describe('annotation utils', () => {
         position: [-10, DEFAULT_LINE_OVERFLOW, -10, 20],
         details: { detailsText: 'foo', headerText: '2' },
         tooltipLinePosition: [-10, DEFAULT_LINE_OVERFLOW, -10, 20],
-      }];
+      },
+      ];
       expect(dimensions).toEqual(expectedDimensions);
     });
 
@@ -876,7 +908,12 @@ describe('annotation utils', () => {
       linePosition,
       Position.Bottom,
     );
-    expect(bottomLineTooltipPosition).toEqual({ xPosition: 1, yPosition: 4, xOffset: 50, yOffset: 100 });
+    expect(bottomLineTooltipPosition).toEqual({
+      xPosition: 1,
+      yPosition: 4,
+      xOffset: 50,
+      yOffset: 100,
+    });
 
     const topLineTooltipPosition = getAnnotationLineTooltipPosition(
       chartRotation,
@@ -890,17 +927,32 @@ describe('annotation utils', () => {
       linePosition,
       Position.Left,
     );
-    expect(leftLineTooltipPosition).toEqual({ xPosition: 1, yPosition: 4, xOffset: 0, yOffset: 50 });
+    expect(leftLineTooltipPosition).toEqual({
+      xPosition: 1,
+      yPosition: 4,
+      xOffset: 0,
+      yOffset: 50,
+    });
 
     const rightLineTooltipPosition = getAnnotationLineTooltipPosition(
       chartRotation,
       linePosition,
       Position.Right,
     );
-    expect(rightLineTooltipPosition).toEqual({ xPosition: 3, yPosition: 4, xOffset: 100, yOffset: 50 });
+    expect(rightLineTooltipPosition).toEqual({
+      xPosition: 3,
+      yPosition: 4,
+      xOffset: 100,
+      yOffset: 50,
+    });
   });
   test('should form the string for the position transform given a TransformPoint', () => {
-    const transformString = toTransformString({ xPosition: 1, yPosition: 4, xOffset: 50, yOffset: 100 });
+    const transformString = toTransformString({
+      xPosition: 1,
+      yPosition: 4,
+      xOffset: 50,
+      yOffset: 100,
+    });
     expect(transformString).toBe('translate(calc(1px - 50%),calc(4px - 100%))');
   });
   test('should get the transform for an annotation line tooltip', () => {
@@ -917,11 +969,13 @@ describe('annotation utils', () => {
   test('should compute the tooltip state for an annotation line', () => {
     const cursorPosition: Point = { x: 1, y: 2 };
 
-    const annotationLines: AnnotationLineProps[] = [{
-      position: [1, 2, 3, 4],
-      details: {},
-      tooltipLinePosition: [1, 2, 3, 4],
-    }];
+    const annotationLines: AnnotationLineProps[] = [
+      {
+        position: [1, 2, 3, 4],
+        details: {},
+        tooltipLinePosition: [1, 2, 3, 4],
+      },
+    ];
     const lineStyle = DEFAULT_ANNOTATION_LINE_STYLE;
     const chartRotation: Rotation = 0;
     const localAxesSpecs = new Map();
@@ -1000,11 +1054,13 @@ describe('annotation utils', () => {
 
     const cursorPosition: Point = { x: 1, y: 2 };
 
-    const annotationLines: AnnotationLineProps[] = [{
-      position: [1, 2, 3, 4],
-      details: {},
-      tooltipLinePosition: [1, 2, 3, 4],
-    }];
+    const annotationLines: AnnotationLineProps[] = [
+      {
+        position: [1, 2, 3, 4],
+        details: {},
+        tooltipLinePosition: [1, 2, 3, 4],
+      },
+    ];
     const chartRotation: Rotation = 0;
     const localAxesSpecs: Map<AxisId, AxisSpec> = new Map();
 
@@ -1116,11 +1172,7 @@ describe('annotation utils', () => {
   test('should get associated axis for an annotation', () => {
     const localAxesSpecs = new Map();
 
-    const noAxis = getAnnotationAxis(
-      localAxesSpecs,
-      groupId,
-      AnnotationDomainTypes.XDomain,
-    );
+    const noAxis = getAnnotationAxis(localAxesSpecs, groupId, AnnotationDomainTypes.XDomain);
     expect(noAxis).toBe(null);
 
     localAxesSpecs.set(horizontalAxisSpec.id, horizontalAxisSpec);