+++ src/state/chart_state.ts
@@ -828,6 +828,7 @@ export class ChartStore {
         bboxCalculator,
         this.chartRotation,
         this.chartTheme.axes,
+        this.chartTheme.scales.barsPadding,
       );
       if (dimensions) {
         this.axesTicksDimensions.set(id, dimensions);
@@ -861,6 +862,7 @@ export class ChartStore {
       this.chartTheme.colors,
       this.chartDimensions,
       this.chartRotation,
+      this.chartTheme.scales.barsPadding,
     );
 
     // tslint:disable-next-line:no-console
@@ -882,6 +884,7 @@ export class ChartStore {
       seriesDomains.yDomain,
       totalBarsInCluster,
       this.legendPosition,
+      this.chartTheme.scales.barsPadding,
     );
     // tslint:disable-next-line:no-console
     // console.log({axisTicksPositions});